@@ -1,6 +1,6 @@
 package com.itxiaox.xview;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 public class ShowViewActivity extends AppCompatActivity {
