@@ -2,9 +2,9 @@ package com.itxiaox.xview.picker.slidedatetimepicker;
 
 import java.util.Date;
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
 
 /**
  * <p>This class contains methods for the library client to create
